Be ready for the upcoming module which helps enigma2 images manage everything in one place:
More information: https://github.com/OpenPLi/openpli-oe-core/pull/888

Keep in mind that it is not my (or our team "Open Vision") duty to support your meta and these are just some of the needed variables which we use, you can change them if you wanted but for the start of the HAL module they are good.

They should not affect your normal builds as HAL will read them.